+++ pydocx/openxml/wordprocessing/level_override.py
@@ -10,6 +10,8 @@ from pydocx.openxml.wordprocessing.level import Level
 
 
 class LevelOverride(XmlModel):
+    XML_TAG = 'lvlOverride'
+
     level_id = XmlAttribute(name='ilvl')
     start_override = XmlChild(name='startOverride', attrname='val')
-    level = XmlChild(type=Level, name='lvl')
+    level = XmlChild(type=Level)