+++ CHANGELOG.rst
@@ -23,6 +23,7 @@
   now used by ``openxml.wordprocessing.styles.Style``
 - Implemented several model classes for Numbering.
 - Added numbering property to the numbering definitions part.
+- XmlModels now define their own tags
 
 **0.6.0**